apps/queue/gateway/requirements/core.txt
pika
httpx
fastapi
python-multipart
uvicorn[standard]

Flask
Flask-PyMongo
